Added change and delete buttons to add_reception page

+++ home/views.py
@@ -377,7 +377,9 @@ def reception_add_page(request: HttpRequest, profile_id: int) -> HttpResponse:
     if request.method == "POST":
         delete_id = request.POST.get('delete_id', None)
         if delete_id is not None:
-            ReceptionNotes.objects.get(pk=delete_id).delete()
+            delete_id = int(delete_id)
+            if delete_id > -1:
+                ReceptionNotes.objects.get(pk=delete_id).delete()
             context.update({ 'form': form_class() })
         else:
             form: ReceptionAddForm = form_class(request.POST)
@@ -396,7 +398,7 @@ def reception_add_page(request: HttpRequest, profile_id: int) -> HttpResponse:
                     commit.visit_number = 1
                 commit.save()
                 # add_log
-
+                
                 # bot
                 try:
                     async_to_sync(bot.bot.send_message)(patient.telegramusers.tg_user_id,
